requirements.txt
urllib3==1.23
psycopg2
numpy==1.16.1
pandas==0.22.0
six==1.10.0
oauth2client
gspread
matplotlib==2.2.2
python-dotenv
google-api-python-client
httplib2
PyDrive
pyasn1==0.4.5
pyasn1-modules
statsmodels
seaborn
sklearn
pathlib
xlrd>=0.9.0
